43. Destroy files forever

+++ utils/rm.py
@@ -23,23 +23,23 @@ def rmtree(path_name):
 
 def rmall(pathname):
     root = Path(pathname)
-    print(f'rmtree | {root=}')
+    print(f'rmall | {root=}')
     pattern = '**/*'
-    print(f'rmtree | {root=}')
 
     file_paths = root.glob(pattern)
     files = [item for item in file_paths if item.is_file()]
     for f in files:
-        print(f'rmtree | unlink {f}')
+        print(f'rmall | unlink {f}')
         f.unlink()
 
     file_paths = root.glob(pattern)
     dirs = [item for item in file_paths if item.is_dir()]
     dirs = sorted(dirs, key=lambda d: str(d.as_posix()).count('/'), reverse=True)
     for d in dirs:
-        print(f'rmtree | rmdir {d}')
+        print(f'rmall | rmdir {d}')
         d.rmdir()
-    root.rmdir()
+    if root.exists():
+        root.rmdir()
 
 
 if __name__ == '__main__':
